Creates DungeonTimeUnit.
Fixes the debug wait integer overflow at multiplication by 1000 bug.
Using two-parameter methods in Date instead of convenience methods.
Reformatted test imports.

+++ src/main/java/org/dungeon/io/Sleeper.java
@@ -29,9 +29,9 @@ public final class Sleeper {
   /**
    * Sleeps for a specified amount of milliseconds.
    *
-   * @param milliseconds how many milliseconds should this thread sleep for.
+   * @param milliseconds how many milliseconds should this thread sleep for
    */
-  public static void sleep(int milliseconds) {
+  public static void sleep(long milliseconds) {
     if (milliseconds > 0) {
       try {
         Thread.sleep(milliseconds);
